Refactored simulation code to be a bit more compartmentalized
